Refactored context to separate concerns
